refactor(conversations): move the reattach family out of the server
`reattach_running_turn/1`, `reap_orphan_sessions/1`,
`attempt_session_attach/4`, `mark_orphan/3` and `find_running_turn/1`
become `Fountain.Conversations.Reattachment`, which already owned
`acp_peer/3` — the ACP half of the same path — plus the runner
reconnect and transport-failure handling. The #1369 subtraction
pattern, following #2037.

The bodies move verbatim. Three edits and no others: two `defp`s
become `def` (the server still calls `reattach_running_turn/1` and
`find_running_turn/1`), and `Reattachment.acp_peer(...)` loses its
now-redundant module prefix. Every function is still a plain call
from the server process, so `self()` in `Managoat.Sandbox.attach/3`
and in the peer still names the server.

163 lines out of `conversation_server.ex`, putting it at 2466.

The three `_unsafe_` reads keep their justification inline rather
than growing `.credo.exs`'s `exempt_paths`, as in #2037: the family
is not a GenServer or a system sweep, it reads the conversation the
server is bound to and the running turn it looked up for it.

Refs #1369, #2032, #1767.
