+++ apps/fountain/lib/fountain/conversations/conversation_server.ex
@@ -1164,7 +1164,7 @@ defmodule Fountain.Conversations.ConversationServer do
               sandbox_started_at: Lifecycle.clock_start(sandbox)
           }
 
-          new_state = reattach_running_turn(%{new_state | current_turn: nil})
+          new_state = Reattachment.reattach_running_turn(%{new_state | current_turn: nil})
 
           new_state =
             Reattachment.finish_runner_reconnect(
@@ -1227,7 +1227,7 @@ defmodule Fountain.Conversations.ConversationServer do
             "transient; sandbox row left untouched"
         )
 
-        running_turn = find_running_turn(state.conversation_id)
+        running_turn = Reattachment.find_running_turn(state.conversation_id)
 
         if sandbox.provider == "runner" and
              reason in [
@@ -1245,169 +1245,6 @@ defmodule Fountain.Conversations.ConversationServer do
           {:stop, :normal, state}
         end
     end
-  end
-
-  # If a turn is marked `running` in the DB and the sprite has an active
-  # detachable session, reattach to it: the WebSocket reconnects, stdout
-  # continues streaming where it left off, and the eventual `:exit` message
-  # closes the turn cleanly. If no active session is found, the command
-  # finished while the BEAM was down — we don't know the exit code, so
-  # mark the orphaned turn `interrupted` so the user gets a clear signal.
-  defp reattach_running_turn(state) do
-    running_turn = find_running_turn(state.conversation_id)
-
-    if is_nil(running_turn) do
-      # No turn was in flight, but a deploy also killed the idle peer that
-      # outlives a turn (#817), leaving its detachable adapter session running
-      # with nothing to drive it. Reap it — by this conversation's tag, never
-      # the head of the list: a co-tenant's live turn on the same machine must
-      # not be touched (ADR 0023, #1058).
-      reap_orphan_sessions(state)
-    else
-      case Managoat.Sandbox.Retry.with_backoff(
-             fn -> Managoat.Sandbox.list_sessions(state.handle) end,
-             label: "session list on reattach"
-           ) do
-        {:ok, sessions} ->
-          # Don't filter by `is_active`: a detached session reports
-          # `is_active: false` while no client is connected, but the
-          # underlying exec is alive and `attach_session` resumes its
-          # stream (replaying the session buffer + live-tailing).
-          #
-          # Match on the conversation tag, never the head of the list: with
-          # several conversations on one machine, the head is as likely to be
-          # someone else's process as ours (`Fountain.Conversations.Identity`).
-          case Fountain.Conversations.Identity.reattach_session(
-                 state.handle,
-                 sessions,
-                 state.conversation_id
-               ) do
-            :none ->
-              mark_orphan(state, running_turn, "no_active_session")
-              state
-
-            {:ok, session, matched_by} ->
-              attempt_session_attach(state, running_turn, session, matched_by)
-          end
-
-        {:error, reason} ->
-          Logger.warning("list_sessions failed during reattach: #{inspect(reason)}")
-          mark_orphan(state, running_turn, "list_sessions_failed")
-          state
-      end
-    end
-  end
-
-  # After a restart with no turn in flight, stop this conversation's own
-  # leftover adapter sessions (#817). Matched by tag; a session tagged for
-  # another conversation, or untagged, is left alone.
-  defp reap_orphan_sessions(state) do
-    case Managoat.Sandbox.list_sessions(state.handle) do
-      {:ok, sessions} ->
-        mine =
-          Fountain.Conversations.Identity.owned_sessions(
-            state.handle,
-            sessions,
-            state.conversation_id
-          )
-
-        Enum.each(mine, &Connection.reap_session(state.handle, &1.id))
-
-        outcome = if mine == [], do: "no_running_turn", else: "orphan_session_reaped"
-
-        Output.publish_stage(state.conversation_id, "reattach", "done", %{
-          outcome: outcome,
-          reaped: length(mine)
-        })
-
-        state
-
-      {:error, _reason} ->
-        Output.publish_stage(state.conversation_id, "reattach", "done", %{
-          outcome: "no_running_turn"
-        })
-
-        state
-    end
-  end
-
-  defp attempt_session_attach(state, running_turn, session, matched_by) do
-    conv = Conversations._unsafe_get_conversation!(state.conversation_id)
-    acp? = Fountain.RuntimeDispatch.acp_enabled?(conv.runtime)
-
-    case Managoat.Sandbox.attach(state.handle, session.id, owner: self(), stdin: true) do
-      {:ok, idle_command} when acp? and is_nil(running_turn.acp_prompt_id) ->
-        # The previous peer died before it wrote `session/prompt` (or the turn
-        # predates the column). The adapter is sitting idle in its handshake
-        # with nothing to answer, and no peer can pick that up: the ids it
-        # would need are gone with the process. Stop it — otherwise it lingers
-        # as a session the next reattach could bind to — and orphan the turn.
-        Managoat.Sandbox.stop_command(idle_command)
-        mark_orphan(state, running_turn, "acp_prompt_not_sent")
-        state
-
-      {:ok, command} ->
-        # sprites replays the tail of the session's buffered output before
-        # live-tailing. On the legacy path, count the bytes we already
-        # persisted for this turn so the stdout/stderr handlers can drop the
-        # replayed prefix. On the ACP path the peer re-encodes protocol lines
-        # so byte counts do not line up; the replayed lines are matched by
-        # content instead (`replay_dedup`).
-        replay_skip =
-          if acp?,
-            do: %{},
-            else:
-              Conversations._unsafe_output_bytes_by_stream(
-                state.conversation_id,
-                running_turn.id
-              )
-
-        Output.publish_stage(state.conversation_id, "reattach", "done", %{
-          outcome: "session_attached",
-          matched_by: matched_by,
-          session_id: session.id,
-          turn_id: running_turn.id,
-          turn_number: running_turn.turn_number,
-          replay_skip_bytes: replay_skip,
-          acp_prompt_id: running_turn.acp_prompt_id
-        })
-
-        {:ok, _} = Conversations.update_conversation(conv, %{status: "running"})
-
-        # turn_metrics stays nil on purpose, so this turn contributes no
-        # duration sample (#536). Its start is in a previous BEAM lifetime:
-        # monotonic time isn't comparable across a restart, and measuring
-        # from the row's started_at would fold the whole deploy gap into the
-        # histogram. A missing sample beats a wrong one.
-        state = %{
-          state
-          | current_command: command,
-            current_command_ref: command.ref,
-            current_turn: running_turn,
-            replay_skip: replay_skip
-        }
-
-        if acp?, do: Reattachment.acp_peer(state, running_turn, conv), else: state
-
-      {:error, reason} ->
-        Logger.warning("attach_session failed: #{inspect(reason)}")
-        mark_orphan(state, running_turn, "attach_failed")
-        state
-    end
-  end
-
-  defp mark_orphan(_state, running_turn, why),
-    do: Conversations._unsafe_orphan_turn(running_turn, why)
-
-  defp find_running_turn(conv_id) do
-    import Ecto.Query
-
-    Fountain.Repo.one(
-      from t in Fountain.Conversations.Turn,
-        where: t.conversation_id == ^conv_id and t.status == "running",
-        order_by: [desc: t.turn_number],
-        limit: 1
-    )
   end
 
   # ── sprite environment and egress (ADR 0019 gate 1a) ──────────────────────
